Replace with_backing_file File initializer with normal method.
Files are must now always be created with the new initializer. Then, a
backing file can be set using set_backing_file on the instance.
This prepares for the introduction of checksum support. As we want to
make this optional, we would have to introduce new "with_checksum"
initializers for each existing initializer. Providing methods on the
instances to enable optional features avoids this combinatoric
explosion.